Added the possibility to create settings
I think it's a good idea to have the possibility,
to change the title and the name of the CMS via a
configuration file. Therefor I created the class Settings
and a configuration file in config/cms/fortress/global_settings.yml.

Furthermore, I moved the Error class into a Cms::Fortress namespace.
I changed the raise and rescue code to the new Error class