make the config for tinymce generic
when one is adding configuration options for tinymce
into config/tinymce.yml, it was only possible to change
existing options in sprocket_helper options. I changed
the behaviour so that it's now possible to add any
oprions into config/tinymce.yml. Every key / value pair will
be added to tinymce.init()

Important: it is not checkd if the options are valid or the syntax
is correct