integrate the Settings mechanism for tinymce_init
We already have the Settings mechanism and I changed the code in tinymce_init to also use the Settings stuff. This requires to put the yml file into /config/cms/fortress in the Rails application using cms-fortress. This could be a breaking change if someone is already using this. The yml file was placed in the root directory of the Rails application using cms-fortress.