extenden tinymce configuration
this extends the configuration possibilities for tinymce. Because we are using a yaml file for all setting, we are kind of restricted to create JavaScript objects for the tinymce configuration. Unfortuanetly. So I added a possibility to add configuration settings like is. The key of the setting in the yaml file has the postfix [plain] like

  style_formats[plain]: |
   [
      {title: "TLN", items: [
        {title: 'Horizontale Linie', inline: 'hr'},
        {title: 'Red text', inline: 'span', styles: {color: '#ff0000'}},
        {title: 'Red header', block: 'h1', styles: {color: '#ff0000'}},
        {title: 'Example 1', inline: 'span', classes: 'example1'}
      ]}
   ]

The code above will be added to tinymce.init({ ... }) like it is.

It's not really nice but we need this! Maybe we should rethink the configuration of tinymce and move it to plain JavaScript.