overwrite comfy admin cms files _page_form
This is an improvement for edit pages in the cms admin, when images are uploaded. I added info of the image, a thumbnail and also the possibility to delete the image. The page is being reloaded after the image has been deleted.

Example is here: https://drive.google.com/open?id=0B2xXhdpvuHcGaGFTYU05ZGZFNFU&authuser=0