tinymce is default editor but there is also redactor
since Comfy 1.12.5, there is also Redactor integrated, which comes with a image and file manager per default. So I extended the gloabl_configuration and added a flag use_tinymce with valu true. If one wants to integrate Redactor, he can simply set this to false and Redactor (if configured) will be loaded